feat: configure docker -it flags dynamically
Configure docker -it flags dynamically according to how the command is being
used.

This is useful to allow wrapped software how to behave depending on
wether they are running in an interactive environment, or a shell
script, or depending on wether their input & output are being piped or
not.

One clear example is output colorization (that might be problematic when
the program output is redirected to a log).

Another example might be interactive wizards, which can be problematic
if the program is executed through an automated script where no human
input is available.