fix: resolve prj dir in a smarter way
Due to a commitizen-related workaround (where we create a temporary
symbolic link to the .git directory), sometimes (after crashes) the
mentioned link is not properly removed.

The dangling symbolic link caused the previous `getProjectDirectory`
function to return a wrong value.

This patch adds code detecting the existence of such symbolic link in
order to address this problem.